refactor(reduced_fingerprint): int type of morgan vectors

+++ src/signature/enumerate_utils.py
@@ -62,7 +62,7 @@ def reduced_fingerprint(smi, radius=2, nbits=2048, useFeatures=False):
         tmp = [(key, info[key][i][1]) for key in info for i in range(len(info[key])) if info[key][i][0] == j]
         indice = np.argmax([tmp[i][1] for i in range(len(tmp))])
         morgan_tmp.append(tmp[indice][0])
-    morgan = np.zeros((nbits,))
+    morgan = np.zeros((nbits,), dtype=int)
     for i in range(morgan.shape[0]):
         if i in morgan_tmp:
             morgan[i] = morgan_tmp.count(i)
